requirements.txt
docker-py
ansible
pywinrm
pywinrm[kerberos]
invocations
invoke
python-terraform
